requirements.txt
# Core
streamlit==1.31.0
python-dotenv==1.0.0

# LLM
anthropic==0.18.1
openai==1.12.0

# RAG
langchain==0.1.10
langchain-community==0.0.24
faiss-cpu==1.7.4
chromadb==0.4.22
sentence-transformers==2.3.1

# Multimodal
pytesseract==0.3.10
Pillow==10.2.0
openai-whisper==20231117
pydub==0.25.1

# Tools
numpy==1.26.4
pandas==2.2.0
sympy==1.12

# Storage
sqlite3